Refactor to avoid the app menu code from getting so complex that swift compiler gets confused whenever you touch the giant nest of { {} } stuff
